requirements.txt
langchain==0.3.13
openai_whisper==20240930
python-dotenv==1.0.1
spacy==3.8.3
whisper==1.1.10
yt_dlp==2024.12.13
